log_analyzer/requirements.txt
langchain_community==0.4.1
langchain_core==1.2.23
langchain_openai==1.1.12
pydantic==2.12.5
python-dotenv==1.2.2
